Attempt Monster, run clippy, attempt to compile.
What have you done?
