+++ Cargo.toml
@@ -17,6 +17,7 @@ maintenance = { status = "experimental" }
 
 [dependencies]
 structopt = "0.2"
+byteorder = "1.2.6"
 
 [dev-dependencies]
 minimp3 = "0.3"
